Move tag_release into a proper package
